+++ clients/team_allocation_component/src/team_allocation/network/queries/getTeamAllocations.ts
@@ -1,4 +1,4 @@
-import { Allocation } from 'src/team_allocation/interfaces/allocation'
+import { Allocation } from '../../interfaces/allocation'
 import { teamAllocationAxiosInstance } from '../teamAllocationServerConfig'
 
 export const getTeamAllocations = async (coursePhaseID: string): Promise<Allocation[]> => {
