+++ clients/team_allocation_component/src/team_allocation/pages/TeamAllocation/TeamAllocationPage.tsx
@@ -1,8 +1,8 @@
 import type React from 'react'
 import { useMemo } from 'react'
 import { useQuery } from '@tanstack/react-query'
-import { useParams, Link } from 'react-router-dom'
-import { Loader2, Users, AlertCircle, CheckCircle2, ArrowRight } from 'lucide-react'
+import { useParams } from 'react-router-dom'
+import { Loader2, Users } from 'lucide-react'
 
 import type { CoursePhaseParticipationsWithResolution, Student } from '@tumaet/prompt-shared-state'
 import type { Team } from '../../interfaces/team'
@@ -14,20 +14,14 @@ import { getTeamAllocations } from '../../network/queries/getTeamAllocations'
 
 import { ErrorPage } from '@/components/ErrorPage'
 import { ManagementPageHeader } from '@/components/ManagementPageHeader'
-import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader } from '@/components/ui/card'
 import { Badge } from '@/components/ui/badge'
 import { Separator } from '@/components/ui/separator'
+import { AllocationSummaryCard } from './components/AllocationSummaryCard'
 
-/**
- * Page that displays the allocated teams for a course phase together with their members.
- */
 export const TeamAllocationPage: React.FC = () => {
   const { phaseId } = useParams<{ phaseId: string }>()
 
-  /**
-   * --- Data Fetching -------------------------------------------------------
-   */
   const {
     data: fetchedTeams,
     isPending: isTeamsPending,
@@ -58,10 +52,6 @@ export const TeamAllocationPage: React.FC = () => {
     queryFn: () => getTeamAllocations(phaseId ?? ''),
   })
 
-  /**
-   * --- Derived Data --------------------------------------------------------
-   * Build a map <courseParticipationID, Student> so we can resolve names fast.
-   */
   const participationMap = useMemo(() => {
     const map = new Map<string, Student>()
     coursePhaseParticipations?.participations.forEach((p) => {
@@ -70,9 +60,6 @@ export const TeamAllocationPage: React.FC = () => {
     return map
   }, [coursePhaseParticipations])
 
-  /**
-   * Combine team, allocation and participations into a render‑friendly shape.
-   */
   const teamsWithMembers = useMemo(() => {
     if (!fetchedTeams) return []
 
@@ -92,54 +79,9 @@ export const TeamAllocationPage: React.FC = () => {
     })
   }, [fetchedTeams, teamAllocations, participationMap])
 
-  /**
-   * Calculate allocation statistics
-   */
-  const allocationStats = useMemo(() => {
-    if (!coursePhaseParticipations || !teamAllocations) {
-      return {
-        totalStudents: 0,
-        assignedStudents: 0,
-        unassignedStudents: 0,
-        isComplete: false,
-        percentComplete: 0,
-      }
-    }
-
-    const totalStudents = coursePhaseParticipations.participations.length
-
-    // Get all assigned student IDs from all allocations
-    const assignedStudentIds = new Set<string>()
-    teamAllocations.forEach((allocation) => {
-      allocation.students.forEach((studentId) => {
-        assignedStudentIds.add(studentId)
-      })
-    })
-
-    const assignedStudents = assignedStudentIds.size
-    const unassignedStudents = totalStudents - assignedStudents
-    const isComplete = unassignedStudents === 0 && totalStudents > 0
-    const percentComplete = totalStudents > 0 ? (assignedStudents / totalStudents) * 100 : 0
-
-    return {
-      totalStudents,
-      assignedStudents,
-      unassignedStudents,
-      isComplete,
-      percentComplete,
-    }
-  }, [coursePhaseParticipations, teamAllocations])
-
-  /**
-   * --- Aggregate Request States -------------------------------------------
-   */
   const isPending = isTeamsPending || isCoursePhaseParticipationsPending || isTeamAllocationsPending
-
   const isError = isTeamsError || isParticipationsError || isTeamAllocationsError
 
-  /**
-   * --- UI ------------------------------------------------------------------
-   */
   if (isPending) {
     return (
       <div className='flex justify-center items-center h-64'>
@@ -163,66 +105,10 @@ export const TeamAllocationPage: React.FC = () => {
   return (
     <div className='space-y-6'>
       <ManagementPageHeader>Team Allocations</ManagementPageHeader>
-
-      {/* Allocation Summary Card */}
-      <Card>
-        <CardHeader className='pb-3'>
-          <div className='flex items-center justify-between'>
-            <h3 className='text-lg font-medium'>Allocation Status</h3>
-            {allocationStats.isComplete ? (
-              <Badge className='bg-green-500 hover:bg-green-600'>
-                <CheckCircle2 className='h-3.5 w-3.5 mr-1' />
-                Complete
-              </Badge>
-            ) : (
-              <Badge variant='outline' className='border-amber-500 text-amber-500'>
-                <AlertCircle className='h-3.5 w-3.5 mr-1' />
-                In Progress
-              </Badge>
-            )}
-          </div>
-        </CardHeader>
-        <CardContent className='pb-4'>
-          <div className='grid grid-cols-1 sm:grid-cols-3 gap-4'>
-            <div className='bg-muted/40 rounded-lg p-4 flex flex-col items-center justify-center text-center'>
-              <div className='rounded-full bg-background p-2 mb-2'>
-                <Users className='h-5 w-5 text-primary' />
-              </div>
-              <p className='text-2xl font-bold'>{allocationStats.totalStudents}</p>
-              <p className='text-sm text-muted-foreground'>Total Students</p>
-            </div>
-
-            <div className='bg-green-50 dark:bg-green-900 rounded-lg p-4 flex flex-col items-center justify-center text-center'>
-              <div className='rounded-full bg-green-100 dark:bg-green-900 p-2 mb-2'>
-                <CheckCircle2 className='h-5 w-5 text-green-600 dark:text-green-400' />
-              </div>
-              <p className='text-2xl font-bold text-green-600 dark:text-green-400'>
-                {allocationStats.assignedStudents}
-              </p>
-              <p className='text-sm text-green-600/80 dark:text-green-400/80'>Assigned</p>
-            </div>
-
-            <div className='bg-red-50 dark:bg-red-800 rounded-lg p-4 flex flex-col items-center justify-center text-center'>
-              <div className='rounded-full bg-red-100 dark:bg-amber-900/30 p-2 mb-2'>
-                <AlertCircle className='h-5 w-5 text-red-600 dark:text-amber-400' />
-              </div>
-              <p className='text-2xl font-bold text-red-600'>
-                {allocationStats.unassignedStudents}
-              </p>
-              <p className='text-sm'>Unassigned</p>
-            </div>
-          </div>
-        </CardContent>
-        <CardContent className='pt-0 pb-4 flex justify-end'>
-          {/* Replace with the correct path for your app */}
-          <Button asChild>
-            <Link to={`/phases/${phaseId}/tease`}>
-              Go to Tease
-              <ArrowRight className='ml-2 h-4 w-4' />
-            </Link>
-          </Button>
-        </CardContent>
-      </Card>
+      <AllocationSummaryCard
+        coursePhaseParticipations={coursePhaseParticipations}
+        teamAllocations={teamAllocations}
+      />
 
       {teamsWithMembers.length === 0 ? (
         <Card className='bg-muted/40'>
@@ -230,17 +116,8 @@ export const TeamAllocationPage: React.FC = () => {
             <div className='rounded-full bg-muted p-3 mb-4'>
               <Users className='h-8 w-8 text-muted-foreground' />
             </div>
-            <h3 className='text-lg font-medium mb-2'>No Teams Allocated Yet</h3>
-            <p className='text-muted-foreground mb-4'>
-              Start by creating teams and allocating students to them.
-            </p>
-            {/* Replace with the correct path for your app */}
-            <Button asChild>
-              <Link to={`/phases/${phaseId}/tease`}>
-                Go to Tease
-                <ArrowRight className='ml-2 h-4 w-4' />
-              </Link>
-            </Button>
+            <h3 className='text-lg font-medium mb-2'>No Teams Created Yet</h3>
+            <p className='text-muted-foreground mb-4'>Start by creating teams.</p>
           </CardContent>
         </Card>
       ) : (
